backend/app/schemas/query.py
from pydantic import BaseModel


class QueryRequest(BaseModel):
    question: str


class QueryResponse(BaseModel):
    question: str
    answer: str
    contexts: list[str]
    response_time_ms: int

